pyproject.toml
[project]
name = "rwkv-gepa-distill"
version = "0.1.0"
description = "Minimal scaffold for API-teacher GEPA prompt optimization and RWKV distillation"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "datasets",
    "gepa",
    "litellm",
    "orjson",
    "python-dotenv",
    "pyyaml",
    "tenacity",
    "tqdm",
]
